add return value to aggregate signature to support immutable aggregator
